--- a/src/GraphQL/Language/AST/AbstractNode.cs
+++ b/src/GraphQL/Language/AST/AbstractNode.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -18,9 +19,12 @@ namespace GraphQL.Language.AST
         public CommentNode CommentNode { get; set; }
 
         /// <inheritdoc/>
+        public SourceLocation SourceLocation { get; set; }
+
+        /// <inheritdoc/>
         public virtual IEnumerable<INode> Children => null;
 
         /// <inheritdoc/>
-        public SourceLocation SourceLocation { get; set; }
+        public virtual void Visit<TState>(Action<INode, TState> action, TState state) { }
     }
 }