--- a/src/GraphQL/Language/AST/Argument.cs
+++ b/src/GraphQL/Language/AST/Argument.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -42,6 +43,9 @@ namespace GraphQL.Language.AST
         {
             get { yield return Value; }
         }
+
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state) => action(Value, state);
 
         /// <inheritdoc />
         public override string ToString() => $"Argument{{name={Name},value={Value}}}";