--- a/src/GraphQL/Language/AST/Arguments.cs
+++ b/src/GraphQL/Language/AST/Arguments.cs
@@ -16,44 +16,41 @@ namespace GraphQL.Language.AST
         /// <inheritdoc/>
         public override IEnumerable<INode> Children => _arguments;
 
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state)
+        {
+            if (_arguments != null)
+            {
+                foreach (var arg in _arguments)
+                    action(arg, state);
+            }
+        }
+
         /// <summary>
         /// Adds an argument node to the list.
         /// </summary>
-        public void Add(Argument arg)
-        {
-            if (arg == null)
-                throw new ArgumentNullException(nameof(arg));
-
-            if (_arguments == null)
-                _arguments = new List<Argument>();
-
-            _arguments.Add(arg);
-        }
+        public void Add(Argument arg) => (_arguments ??= new List<Argument>()).Add(arg ?? throw new ArgumentNullException(nameof(arg)));
 
         /// <summary>
         /// Returns the value of an argument node, searching the list of argument nodes by the name of the argument.
         /// </summary>
         public IValue ValueFor(string name)
         {
-            if (_arguments == null)
-                return null;
-
             // DO NOT USE LINQ ON HOT PATH
-            foreach (var x in _arguments)
-                if (x.Name == name)
-                    return x.Value;
+            if (_arguments != null)
+            {
+                foreach (var x in _arguments)
+                {
+                    if (x.Name == name)
+                        return x.Value;
+                }
+            }
 
             return null;
         }
 
         /// <inheritdoc cref="IEnumerable.GetEnumerator"/>
-        public IEnumerator<Argument> GetEnumerator()
-        {
-            if (_arguments == null)
-                return System.Linq.Enumerable.Empty<Argument>().GetEnumerator();
-
-            return _arguments.GetEnumerator();
-        }
+        public IEnumerator<Argument> GetEnumerator() => (_arguments ?? System.Linq.Enumerable.Empty<Argument>()).GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();