--- a/src/GraphQL/Language/AST/FragmentDefinition.cs
+++ b/src/GraphQL/Language/AST/FragmentDefinition.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -47,14 +48,19 @@ namespace GraphQL.Language.AST
 
                 if (Directives != null)
                 {
-                    foreach (var directive in Directives)
-                    {
-                        yield return directive;
-                    }
+                    yield return Directives;
                 }
 
                 yield return SelectionSet;
             }
+        }
+
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state)
+        {
+            action(Type, state);
+            action(Directives, state);
+            action(SelectionSet, state);
         }
 
         /// <inheritdoc/>