--- a/src/GraphQL/Language/AST/Fragments.cs
+++ b/src/GraphQL/Language/AST/Fragments.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace GraphQL.Language.AST
 {
@@ -9,15 +10,12 @@ namespace GraphQL.Language.AST
     /// </summary>
     public class Fragments : IEnumerable<FragmentDefinition>
     {
-        private readonly List<FragmentDefinition> _fragments = new List<FragmentDefinition>();
+        private List<FragmentDefinition> _fragments;
 
         /// <summary>
         /// Adds a fragment definition node to the list.
         /// </summary>
-        public void Add(FragmentDefinition fragment)
-        {
-            _fragments.Add(fragment ?? throw new ArgumentNullException(nameof(fragment)));
-        }
+        public void Add(FragmentDefinition fragment) => (_fragments ??= new List<FragmentDefinition>()).Add(fragment ?? throw new ArgumentNullException(nameof(fragment)));
 
         /// <summary>
         /// Returns the number of fragment definition nodes in the list.
@@ -30,18 +28,20 @@ namespace GraphQL.Language.AST
         public FragmentDefinition FindDefinition(string name)
         {
             // DO NOT USE LINQ ON HOT PATH
-            foreach (var f in _fragments)
-                if (f.Name == name)
-                    return f;
+            if (_fragments != null)
+            {
+                foreach (var f in _fragments)
+                {
+                    if (f.Name == name)
+                        return f;
+                }
+            }
 
             return null;
         }
 
         /// <inheritdoc/>
-        public IEnumerator<FragmentDefinition> GetEnumerator()
-        {
-            return _fragments.GetEnumerator();
-        }
+        public IEnumerator<FragmentDefinition> GetEnumerator() => (_fragments ?? Enumerable.Empty<FragmentDefinition>()).GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
     }