--- a/src/GraphQL/Language/AST/INode.cs
+++ b/src/GraphQL/Language/AST/INode.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -8,13 +9,21 @@ namespace GraphQL.Language.AST
     public interface INode
     {
         /// <summary>
+        /// Returns the node's location within the source document.
+        /// </summary>
+        SourceLocation SourceLocation { get; }
+
+        /// <summary>
         /// Returns a list of children nodes. If the node doesn't have children, returns <see langword="null"/>.
         /// </summary>
         IEnumerable<INode> Children { get; }
 
         /// <summary>
-        /// Returns the node's location within the source document.
+        /// Visits every child node with the specified delegate and state. If the node doesn't have children, does nothing.
         /// </summary>
-        SourceLocation SourceLocation { get; }
+        /// <typeparam name="TState">Type of the provided state.</typeparam>
+        /// <param name="action">Delegate to execute on every child node of this node.</param>
+        /// <param name="state">An arbitrary state passed by the caller.</param>
+        void Visit<TState>(Action<INode, TState> action, TState state);
     }
 }