--- a/src/GraphQL/Language/AST/NameNode.cs
+++ b/src/GraphQL/Language/AST/NameNode.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -30,9 +31,12 @@ namespace GraphQL.Language.AST
         /// </summary>
         public string Name { get; }
 
+        /// <inheritdoc/>
+        public SourceLocation SourceLocation { get; }
+
         IEnumerable<INode> INode.Children => null;
 
         /// <inheritdoc/>
-        public SourceLocation SourceLocation { get; }
+        public void Visit<TState>(Action<INode, TState> action, TState state) { }
     }
 }