--- a/src/GraphQL/Language/AST/Operations.cs
+++ b/src/GraphQL/Language/AST/Operations.cs
@@ -10,39 +10,39 @@ namespace GraphQL.Language.AST
     /// </summary>
     public class Operations : IEnumerable<Operation>
     {
-        private readonly List<Operation> _operations = new List<Operation>();
+        private List<Operation> _operations;
 
         /// <summary>
         /// Returns the number of operation nodes the list contains.
         /// </summary>
-        public int Count => _operations.Count;
+        public int Count => _operations?.Count ?? 0;
 
         /// <summary>
         /// Adds an operation node to the list.
         /// </summary>
-        /// <param name="operation"></param>
-        public void Add(Operation operation)
-        {
-            _operations.Add(operation ?? throw new ArgumentNullException(nameof(operation)));
-        }
+        public void Add(Operation operation) => (_operations ??= new List<Operation>()).Add(operation ?? throw new ArgumentNullException(nameof(operation)));
 
         /// <summary>
         /// Returns the first operation in the list that matches the specified name, or <see langword="null"/> if none are matched.
         /// </summary>
         public Operation WithName(string operationName)
         {
-            return _operations.FirstOrDefault(op => op.Name == operationName);
+            // DO NOT USE LINQ ON HOT PATH
+            if (_operations != null)
+            {
+                foreach (var op in _operations)
+                {
+                    if (op.Name == operationName)
+                        return op;
+                }
+            }
+
+            return null;
         }
 
         /// <inheritdoc/>
-        public IEnumerator<Operation> GetEnumerator()
-        {
-            return _operations.GetEnumerator();
-        }
+        public IEnumerator<Operation> GetEnumerator() => (_operations ?? Enumerable.Empty<Operation>()).GetEnumerator();
 
-        IEnumerator IEnumerable.GetEnumerator()
-        {
-            return GetEnumerator();
-        }
+        IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
     }
 }