--- a/src/GraphQL/Language/AST/ValueNodes/ListValue.cs
+++ b/src/GraphQL/Language/AST/ValueNodes/ListValue.cs
@@ -14,21 +14,47 @@ namespace GraphQL.Language.AST
         /// </summary>
         public ListValue(IEnumerable<IValue> values)
         {
-            Values = values ?? Array.Empty<IValue>();
+            ValuesList = (values ?? throw new ArgumentNullException(nameof(values))).ToList();
+        }
+
+        /// <summary>
+        /// Initializes a new instance with the specified list of values.
+        /// </summary>
+        public ListValue(List<IValue> values)
+        {
+            ValuesList = values ?? throw new ArgumentNullException(nameof(values));
         }
 
         /// <summary>
         /// Returns a <see cref="List{T}">List&lt;object&gt;</see> containing the values of the list.
         /// </summary>
-        public object Value => Values.Select(x => x.Value).ToList();
+        public object Value
+        {
+            get
+            {
+                var list = new List<object>(ValuesList.Count);
+                foreach (var item in ValuesList)
+                    list.Add(item.Value);
+                return list;
+            }
+        }
 
         /// <summary>
         /// Returns a list of the child value nodes.
         /// </summary>
-        public IEnumerable<IValue> Values { get; }
+        public IEnumerable<IValue> Values => ValuesList;
+
+        internal List<IValue> ValuesList { get; private set; }
 
         /// <inheritdoc/>
-        public override IEnumerable<INode> Children => Values;
+        public override IEnumerable<INode> Children => ValuesList;
+
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state)
+        {
+            foreach (var value in ValuesList)
+                action(value, state);
+        }
 
         /// <inheritdoc/>
         public override string ToString()