--- a/src/GraphQL/Language/AST/ValueNodes/ObjectField.cs
+++ b/src/GraphQL/Language/AST/ValueNodes/ObjectField.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 
 namespace GraphQL.Language.AST
@@ -45,6 +46,9 @@ namespace GraphQL.Language.AST
         {
             get { yield return Value; }
         }
+
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state) => action(Value, state);
 
         /// <inheritdoc/>
         public override string ToString() => $"ObjectField{{name='{Name}', value={Value}}}";