--- a/src/GraphQL/Language/AST/ValueNodes/ObjectValue.cs
+++ b/src/GraphQL/Language/AST/ValueNodes/ObjectValue.cs
@@ -14,10 +14,15 @@ namespace GraphQL.Language.AST
         /// </summary>
         public ObjectValue(IEnumerable<ObjectField> fields)
         {
-            if (fields == null)
-                ObjectFields = Array.Empty<ObjectField>();
-            else
-                ObjectFields = fields;
+            ObjectFieldsList = (fields ?? throw new ArgumentNullException(nameof(fields))).ToList();
+        }
+
+        /// <summary>
+        /// Initializes a new instance that contains the specified field nodes.
+        /// </summary>
+        public ObjectValue(List<ObjectField> fields)
+        {
+            ObjectFieldsList = fields ?? throw new ArgumentNullException(nameof(fields));
         }
 
         /// <summary>
@@ -37,22 +42,47 @@ namespace GraphQL.Language.AST
         /// <summary>
         /// Returns the field value nodes that are contained within this object value node.
         /// </summary>
-        public IEnumerable<ObjectField> ObjectFields { get; }
+        public IEnumerable<ObjectField> ObjectFields => ObjectFieldsList;
+
+        internal List<ObjectField> ObjectFieldsList { get; private set; }
 
         /// <summary>
         /// Returns a list of the names of the fields specified for this object value node.
         /// </summary>
-        public IEnumerable<string> FieldNames => ObjectFields.Select(x => x.Name).ToList();
+        public IEnumerable<string> FieldNames
+        {
+            get
+            {
+                var list = new List<string>(ObjectFieldsList.Count);
+                foreach (var item in ObjectFieldsList)
+                    list.Add(item.Name);
+                return list;
+            }
+        }
 
         /// <inheritdoc/>
-        public override IEnumerable<INode> Children => ObjectFields;
+        public override IEnumerable<INode> Children => ObjectFieldsList;
+
+        /// <inheritdoc/>
+        public override void Visit<TState>(Action<INode, TState> action, TState state)
+        {
+            foreach (var field in ObjectFieldsList)
+                action(field, state);
+        }
 
         /// <summary>
         /// Returns the first matching field node contained within this object value node that matches the specified name, or <see langword="null"/> otherwise.
         /// </summary>
         public ObjectField Field(string name)
         {
-            return ObjectFields.FirstOrDefault(x => x.Name == name);
+            // DO NOT USE LINQ ON HOT PATH
+            foreach (var field in ObjectFieldsList)
+            {
+                if (field.Name == name)
+                    return field;
+            }
+
+            return null;
         }
 
         /// <inheritdoc/>