--- a/src/GraphQL/Language/AST/VariableDefinitions.cs
+++ b/src/GraphQL/Language/AST/VariableDefinitions.cs
@@ -10,34 +10,19 @@ namespace GraphQL.Language.AST
     /// </summary>
     public class VariableDefinitions : IEnumerable<VariableDefinition>
     {
-        private List<VariableDefinition> _variables;
+        internal List<VariableDefinition> VariablesList { get; private set; }
 
         /// <summary>
         /// Adds a variable definition node to the list.
         /// </summary>
-        public void Add(VariableDefinition variable)
-        {
-            if (variable == null)
-                throw new ArgumentNullException(nameof(variable));
-
-            if (_variables == null)
-                _variables = new List<VariableDefinition>();
-
-            _variables.Add(variable);
-        }
+        public void Add(VariableDefinition variable) => (VariablesList ??= new List<VariableDefinition>()).Add(variable ?? throw new ArgumentNullException(nameof(variable)));
 
         /// <inheritdoc/>
-        public IEnumerator<VariableDefinition> GetEnumerator()
-        {
-            if (_variables == null)
-                return Enumerable.Empty<VariableDefinition>().GetEnumerator();
-
-            return _variables.GetEnumerator();
-        }
+        public IEnumerator<VariableDefinition> GetEnumerator() => (VariablesList ?? Enumerable.Empty<VariableDefinition>()).GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
 
         /// <inheritdoc/>
-        public override string ToString() => _variables?.Count > 0 ? $"VariableDefinitions{{{string.Join(", ", _variables)}}}" : "VariableDefinitions(Empty)";
+        public override string ToString() => VariablesList?.Count > 0 ? $"VariableDefinitions{{{string.Join(", ", VariablesList)}}}" : "VariableDefinitions(Empty)";
     }
 }