… docs2/site/docs/migrations/migration4.md
… src/GraphQL.ApiTests/GraphQL.approved.txt
… src/GraphQL/Types/QueryArguments.cs
… src/GraphQL/Types/Scalars/EnumerationGraphType.cs
… src/GraphQL/Utilities/Federation/FederatedSchemaBuilder.cs
… src/GraphQL/Validation/BasicVisitor.cs
… src/GraphQL/Validation/DocumentValidator.cs
… src/GraphQL/Validation/Rules/ProvidedNonNullArguments.cs
… src/GraphQL/Validation/ValidationContext.cs