``` ini

BenchmarkDotNet=v0.12.1, OS=Windows 10.0.16299.726 (1709/FallCreatorsUpdate/Redstone3)
Intel Core i7 CPU 920 2.67GHz (Nehalem), 1 CPU, 8 logical and 4 physical cores
Frequency=2607421 Hz, Resolution=383.5207 ns, Timer=TSC
.NET Core SDK=5.0.101
  [Host]     : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT
  DefaultJob : .NET Core 3.1.10 (CoreCLR 4.700.20.51601, CoreFX 4.700.20.51901), X64 RyuJIT

```
before/after

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 446.98 μs | 0.481 μs | 0.402 μs | 4.3945 |     - |     - |  19.59 KB |
|     Fragments | 204.92 μs | 0.048 μs | 0.040 μs | 2.9297 |     - |     - |  12.31 KB |
|          Hero |  30.26 μs | 0.018 μs | 0.015 μs | 1.1902 |     - |     - |   4.95 KB |

|        Method |      Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|-------------- |----------:|---------:|---------:|-------:|------:|------:|----------:|
| Introspection | 426.06 μs | 0.588 μs | 0.491 μs | 3.4180 |     - |     - |  14.16 KB |
|     Fragments | 193.79 μs | 0.123 μs | 0.115 μs | 2.4414 |     - |     - |  10.07 KB |
|          Hero |  29.66 μs | 0.055 μs | 0.051 μs | 1.0681 |     - |     - |   4.48 KB |

My main goal was to reduce the number of heap allocations. It's nice that the speed has also increased. The PR contains a lot of minor fixes, but the main idea is to stop visitors from allocating memory when iterating over children. This was achieved 100%.