@sungam3r 2021-01-08
I don't remember, but I have a feeling that we have reduced memory consumption at the validation stage by x2.